flags, unpack fifo mods, front panel leds
fifo blocks in unpack modified to not allow buffer overflows. Packet
sizes are restricted to a maximum and the packet discarded if this is
exceeded. A maximum number of buffered packets (an optional power of 2
instead of 2 as before) is also enforced. Whole packets are discarded.
This should allow for graceful recovery from some errors without having
 to do a hard reset.

flags are captured using a 'sampler' block. This block allocates space
in a buffer for a flag. After a specified block of data is completed
an index increases to the next space in the buffer. Another process
reads these flags out. This allows flags associated with a certain section
a data pipeline to be recorded, and read out for use in a later pipeline
section. The following output flags are currently supported;
* ADC overflow
* FFT overflow (although not aligned with data)
* Missing packet bit (should use packet discard line from unpacker)
* Timestamp
TODO
* QDR
* Noise diode
* checksum

Front panel leds show some status info

TVG allowing PFB testing added