sync_gen test model
sync_gen copied from RTS and improved to guarantee sync happens
at beginning of timestamp. A race condition previously existed
where forcing a resync could cause the sync to happen incorrectly
relative to the timestamp. This would cause data to be tagged with
the wrong time by up to a few thousand samples.