local_time update, flag_epoch registers
A trigger to capture local time now captures it at all points in the
pipeline. This can then be read for debugging and sanity checking
purposes.

flag_epoch_msw and flag_epoch_lsw added to capture the sync time
as seen just before the pack block. This should be the same as sync_time_msw
and sync_time_lsw.

These changes were made as an fengine was producing the incorrect time but
it was not obvious why...