added debug for resync phase jumps
We have noticed that the phase (delay) of the fengines change relative to each other when they are reset. We suspect the CD logic that forces a zero delay. This design has a snapshot block compiled-in around the CD write control block to help debug.

Also, logic fix to the sync_timeout logic that will force a resync if the sync pulse doesn't pop out the other end of the pipeline after a defined wait period (2**30 clocks).