CD reduce reorder offset from 256 to 128 words
The CD reorder used to try to keep the read pointer halfway behind the write pointer. But sometimes the one leapfrogged the other and things would go out of sync. Given that we're only ever reading and writing linearly, the output is only ever out of order by a few words, so I've reduced the read pointer to only 1/4 behind the write pointer.