Wideband Fengs delay tracking and CT status regs
Delay tracking fixes for delay and phase rate terms, which were out by a factor
of two.

No change to the underlying CT functionality, but have updated the status
registers for the CT.

To reduce the number of required workspace variables, xengbits workspace
register has been replaced with n_bits_xengs, as per xengine. This should
be set to 8 for MeerKAT compiles.

You need the following workspace variables to compile MeerKAT F and X engines:
 * n_bits_ants
 * n_bits_xengs
 * fft_stages