time debug
time can be injected into the data stream. This time is compared
with locally regenerated time before the PFB, before the coarse
delay block.

The time difference between ingress and PFB, coarse delay and packetiser
block is also calculated and can be compared across F-engines to look
for significant differences.