<fix> beam steering coeff change
Beam steering coefficients were stored in Distributed RAM loaded
indirectly via registers. This is too slow, so Shared BRAM now store
the coefficients. This is not as efficient in terms of resource use
and possibly timing, but should allow faster loading as a bulk transfer
of the coefficients can be performed.

A snap block is also included to aid with software development.

Makes timing but untested.