1k f-engine with configurable watchdo timeout
F-engines expect packets to appear regularly from Digitisers. A watchdog
timer in the pkt_reorder module (in unpack subsytem) raises an error and
resyncs to the datastream, causing potential packet loss. The new MeerKAT+
software DSim seems to release packets in a more bursty way so our f-engines
aren't happy.

This build has the default timeout (2^11 FPGA clock cycles) but has a
register (unpack_timeoutperiod) that allows this to be modified.

The fpg is untested.