SKARAB DSim packet length fix
The dvalid input signal on the 40GbE core is actually 4 separate inputs,
1 for each 64 bit component. However, this is combined into a single
input preventing us from distinguising between these. The DSim packet
does not cleanly divide into a 256 bit word so the packet length is
generated incorrectly in the core as it is rounded up. When integrating
for MeerKAT+, the SPEAD2 receiver was rejecting the packets as they were
the incorrect length.

The fix was to modify the yellow block so that the 4 signals could be
carried through, and generate these correctly. A fair amount of logic
along the path also needed modification to work with this 4 component
bus that was 1.

fpg untested.